pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aion-daemon"
version = "0.1.0"
description = "🏛️ AION: The Autonomous Architect"
dependencies = [
    "watchdog>=2.3.1",
    "GitPython>=3.1.30",
    "SQLAlchemy>=2.0.23",
    "alembic>=1.13.0",
    "typer[all]>=0.9.0",
    "ollama>=0.1.0",
    "mcp>=0.1.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "llama-index-core",
    "llama-index-readers-file",
    "llama-index-embeddings-huggingface",
    "striprtf"
]

[project.scripts]
aion-daemon = "aion.core.heartbeat:start_daemon_main"
aion-cli = "aion.interface.cli:app"